Move the HealthChecks in the "controllers" package
