+++ admin-jobs/app/conf/HealthCheck.scala
@@ -1,9 +1,7 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(
-  9015,
-  "/news-alert/alerts"
-)
+import controllers.HealthCheck
+
 
 trait AdminJobsHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
