+++ admin-jobs/test/AdminJobsTestSuite.scala
@@ -1,4 +1,5 @@
 import controllers.BreakingNews.{BreakingNewsUpdaterTest, BreakingNewsApiTest}
+import controllers.HealthCheck
 import org.scalatest.Suites
 import test.SingleServerSuite
 
@@ -8,5 +9,5 @@ class AdminJobsTestSuite extends Suites (
   new controllers.NewsAlertControllerTest
 )
 with SingleServerSuite {
-  override lazy val port: Int = conf.HealthCheck.testPort
+  override lazy val port: Int = HealthCheck.testPort
 }
