+++ archive/app/conf/context.scala
@@ -1,6 +1,6 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(9003, "/404/www.theguardian.com/Adzip/adzip-fb.html")
+import controllers.HealthCheck
 
 trait ArchiveHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
