+++ article/app/conf/context.scala
@@ -1,6 +1,6 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(9004, "/world/2012/sep/11/barcelona-march-catalan-independence")
+import controllers.HealthCheck
 
 trait ArticleHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
