+++ article/conf/routes
@@ -5,9 +5,9 @@
 # For dev machines
 GET     /assets/*path               dev.DevAssetsController.at(path)
 
-GET     /_healthcheck               conf.HealthCheck.healthCheck()
+GET     /_healthcheck               controllers.HealthCheck.healthCheck()
 
-GET     /_cdn_healthcheck           conf.HealthCheck.healthCheck()
+GET     /_cdn_healthcheck           controllers.HealthCheck.healthCheck()
 
 # Newspaper pages paths
 # gallery format (?)
