+++ article/test/CdnHealthCheckTest.scala
@@ -1,5 +1,6 @@
 package test
 
+import controllers.HealthCheck
 import org.scalatest.{DoNotDiscover, FlatSpec, Matchers}
 import play.api.test.Helpers._
 import org.scalatest.concurrent.ScalaFutures
@@ -7,7 +8,7 @@ import org.scalatest.concurrent.ScalaFutures
 @DoNotDiscover class CdnHealthCheckTest extends FlatSpec with Matchers with ConfiguredTestSuite with ScalaFutures {
 
   "CDN health check" should "mimic the instance health check" in {
-    val controller = conf.HealthCheck
+    val controller = HealthCheck
     // Cache internal healthCheck results before to test endpoints
     whenReady(controller.runChecks) { _ =>
       status(controller.healthCheck()(TestRequest("/_healthcheck"))) should be (200)
