+++ article/test/package.scala
@@ -1,5 +1,6 @@
 package test
 
+import controllers.HealthCheck
 import org.scalatest.{Suites, Tag}
 
 object ArticleComponents extends Tag("article components")
@@ -15,5 +16,5 @@ class ArticleTestSuite extends Suites (
   new PublicationControllerTest
 ) with SingleServerSuite {
 
-  override lazy val port: Int = conf.HealthCheck.testPort
+  override lazy val port: Int = HealthCheck.testPort
 }
