+++ commercial/app/controllers/HealthCheck.scala
@@ -1,0 +1,13 @@
+package controllers
+
+import conf.AnyGoodCachedHealthCheck
+
+object HealthCheck extends AnyGoodCachedHealthCheck(
+  9005,
+  "/commercial/soulmates/mixed.json",
+  "/commercial/masterclasses.json",
+  "/commercial/travel/offers.json",
+  "/commercial/jobs.json",
+  "/commercial/money/bestbuys.json",
+  "/commercial/books/books.json"
+)
