+++ commercial/conf/routes
@@ -5,7 +5,7 @@
 # For dev machines
 GET        /assets/*path                                                        dev.DevAssetsController.at(path)
 GET        /commercial/test-page                                                controllers.commercial.CreativeTestPage.allComponents(k: List[String])
-GET        /_healthcheck                                                        conf.HealthCheck.healthCheck()
+GET        /_healthcheck                                                        controllers.HealthCheck.healthCheck()
 
 # Travel offer merchandising components
 GET        /commercial/travel/offers.json                                       controllers.commercial.TravelOffersController.renderTravel
