+++ commercial/test/test/CommercialTestSuite.scala
@@ -1,5 +1,6 @@
 package test
 
+import controllers.HealthCheck
 import model.commercial._
 import model.commercial.books._
 import org.scalatest.Suites
@@ -20,5 +21,5 @@ class CommercialTestSuite extends Suites (
   new BookTest
   ) with SingleServerSuite {
 
-  override lazy val port: Int = conf.HealthCheck.testPort
+  override lazy val port: Int = HealthCheck.testPort
 }
