+++ diagnostics/app/conf/context.scala
@@ -1,6 +1,7 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(9006, "/robots.txt")
+import controllers.HealthCheck
+
 
 trait DiagnosticsHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
