+++ diagnostics/app/controllers/HealthCheck.scala
@@ -1,0 +1,5 @@
+package controllers
+
+import conf.AllGoodCachedHealthCheck
+
+object HealthCheck extends AllGoodCachedHealthCheck(9006, "/robots.txt")
