+++ discussion/app/conf/context.scala
@@ -1,6 +1,7 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(9007, "/discussion/p/37v3a")
+import controllers.HealthCheck
+
 
 trait DiscussionHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
