+++ discussion/conf/routes
@@ -5,7 +5,7 @@
 # For assets on dev machines
 GET         /assets/*path                                                       dev.DevAssetsController.at(path)
 
-GET         /_healthcheck                                                       conf.HealthCheck.healthCheck()
+GET         /_healthcheck                                                       controllers.HealthCheck.healthCheck()
 
 GET         /discussion/comment-counts                                          controllers.CommentCountController.commentCount(shortUrls: String)
 GET         /discussion/comment-counts.json                                     controllers.CommentCountController.commentCountJson(shortUrls: String)
