+++ discussion/test/package.scala
@@ -1,6 +1,7 @@
 package test
 
 import conf.Configuration
+import controllers.HealthCheck
 import org.scalatest.Suites
 import play.api.libs.ws.ning.NingWSResponse
 import recorder.HttpRecorder
@@ -82,7 +83,7 @@ class DiscussionTestSuite extends Suites (
   new CommentCountControllerTest,
   new ProfileTest
   ) with SingleServerSuite {
-  override lazy val port: Int = conf.HealthCheck.testPort
+  override lazy val port: Int = HealthCheck.testPort
 
   // Inject stub api.
   controllers.delegate.api = new DiscussionApiStub()
