+++ facia-press/app/controllers/HealthCheck.scala
@@ -1,8 +1,10 @@
-package conf
+package controllers
 
+import _root_.play.api.mvc.{Action, Results}
+import conf.HealthCheckController
 import frontpress.ToolPressQueueWorker
 import org.joda.time.DateTime
-import _root_.play.api.mvc.{Results, Action}
+
 import scala.concurrent.Future.successful
 
 object HealthCheck extends HealthCheckController with Results {
