+++ facia-press/conf/routes
@@ -1,4 +1,4 @@
-GET        /_healthcheck        conf.HealthCheck.healthCheck()
+GET        /_healthcheck        controllers.HealthCheck.healthCheck()
 GET        /                    controllers.Application.index
 GET        /debug/config        controllers.Application.showCurrentConfig
