+++ facia/conf/routes
@@ -6,8 +6,8 @@
 # For dev machines
 GET        /assets/*path                                                            dev.DevAssetsController.at(path)
 
-GET        /_healthcheck                                                            conf.HealthCheck.healthCheck()
-GET        /_fronts_cdn_healthcheck                                                 conf.HealthCheck.healthCheck()
+GET        /_healthcheck                                                            controllers.HealthCheck.healthCheck()
+GET        /_fronts_cdn_healthcheck                                                 controllers.HealthCheck.healthCheck()
 
 GET        /_agentcontents                                                          controllers.FaciaController.renderAgentContents
