+++ identity/app/conf/Management.scala
@@ -1,6 +1,7 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(9010, "/signin")
+import controllers.HealthCheck
+
 
 trait IdentityHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
