+++ onward/app/conf/context.scala
@@ -1,10 +1,6 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(
-  9011,
-  "/top-stories.json",
-  "/most-read/society.json"
-)
+import controllers.HealthCheck
 
 trait OnwardHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
