+++ onward/app/controllers/HealthCheck.scala
@@ -1,0 +1,9 @@
+package controllers
+
+import conf.AllGoodCachedHealthCheck
+
+object HealthCheck extends AllGoodCachedHealthCheck(
+  9011,
+  "/top-stories.json",
+  "/most-read/society.json"
+)
