+++ rss/app/conf/HealthCheck.scala
@@ -1,9 +1,7 @@
 package conf
 
-object HealthCheck extends AllGoodCachedHealthCheck(
-  9014,
-  "/books/harrypotter/rss"
-)
+import controllers.HealthCheck
+
 
 trait RssHealthCheckLifeCycle extends CachedHealthCheckLifeCycle {
   override val healthCheckController = HealthCheck
