+++ rss/app/controllers/HealthCheck.scala
@@ -1,0 +1,8 @@
+package controllers
+
+import conf.AllGoodCachedHealthCheck
+
+object HealthCheck extends AllGoodCachedHealthCheck(
+  9014,
+  "/books/harrypotter/rss"
+)
