+++ rss/conf/routes
@@ -3,7 +3,7 @@
 # ~~~~
 
 
-GET        /_healthcheck                                               conf.HealthCheck.healthCheck()
+GET        /_healthcheck                           controllers.HealthCheck.healthCheck()
 
 GET        /$leftSide<[^+]+>+*rightSide/rss                            controllers.RssController.renderCombinerRss(leftSide, rightSide)
 GET        /*path/rss                                                  controllers.RssController.renderRss(path)
