+++ sport/app/football/controllers/HealthCheck.scala
@@ -1,0 +1,9 @@
+package football.controllers
+
+import conf.AllGoodCachedHealthCheck
+
+object HealthCheck extends AllGoodCachedHealthCheck(
+  9013,
+  "/football/live",
+  "/football/premierleague/results"
+)
