+++ sport/conf/routes
@@ -5,7 +5,7 @@
 # For dev machines
 GET        /assets/*path                                                                       dev.DevAssetsController.at(path)
 
-GET        /_healthcheck                                                                       conf.HealthCheck.healthCheck()
+GET        /_healthcheck                                                                       football.controllers.HealthCheck.healthCheck()
 
 GET        /sport/cricket/match/:matchDate/:teamId.json                                        controllers.CricketMatchController.renderMatchIdJson(matchDate, teamId)
 GET        /sport/cricket/match/:matchDate/:teamId                                             controllers.CricketMatchController.renderMatchId(matchDate, teamId)
